% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{als_list}
\alias{als_list}
\title{List ASCII data example}
\format{
A data frame with 200 rows and 38 columns:
}
\source{
<https://www.fda.gov/drugs/questions-and-answers-fdas-adverse-event-reporting-system-faers/fda-adverse-event-reporting-system-faers-latest-quarterly-data-files>
}
\usage{
als_list
}
\description{
A list containing data from FDA website. The list only contains safety reports which the ADR primary suspect drug was indicated for ALS.
List originated from \link{retrieve_faersascii}
}
\keyword{datasets}
