% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_ascii.R
\name{arrange_date}
\alias{arrange_date}
\title{Convert a date string into a date format}
\usage{
arrange_date(date_string)
}
\arguments{
\item{date_string}{A string vector with multiple formats (8, 6 or 4 digits)}
}
\value{
A converted Date
}
\description{
Convert a date string into a date format
}
\examples{
arrange_date("2020")
arrange_date("202006")
arrange_date("20200601")
}
