% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{estimate_chisq}
\alias{estimate_chisq}
\title{Estimate Chi-Squared test with yates correction}
\usage{
estimate_chisq(n11, n10, n01, n00)
}
\arguments{
\item{n11}{Number of events of interest within the group of interest}

\item{n10}{Number of events of interest from all groups}

\item{n01}{Number of all events within the group of interest}

\item{n00}{Number of all events from all groups}
}
\value{
list with Chi-squared statistic and p-value
}
\description{
Estimate Chi-Squared test with yates correction
}
\examples{
estimate_chisq(n11 = 20, n10 = 10, n01 = 200, n00 = 200)
}
