% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{estimate_ror}
\alias{estimate_ror}
\title{Estimate Reporting Odds Ratio}
\usage{
estimate_ror(n11, n10, n01, n00, ic_range = 0.95)
}
\arguments{
\item{n11}{Number of events of interest within the group of interest}

\item{n10}{Number of events of interest from all groups}

\item{n01}{Number of all events within the group of interest}

\item{n00}{Number of all events from all groups}

\item{ic_range}{Confidence Interval range}
}
\value{
list with ROR estimate and a vector with the IC boundaries
}
\description{
Estimate Reporting Odds Ratio
}
\examples{
estimate_ror(n11 = 20, n10 = 10, n01 = 200, n00 = 200, ic_range = 0.90)
}
