% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_xml.R
\name{faersxml_to_r}
\alias{faersxml_to_r}
\title{Convert FAERS xml to an R list}
\usage{
faersxml_to_r(xml_address)
}
\arguments{
\item{xml_address}{XML address file}
}
\value{
a list containing all the elements from `xml_address`
}
\description{
Convert FAERS xml to an R list
}
