% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_ascii.R
\name{get_duplicate_caseids}
\alias{get_duplicate_caseids}
\title{Get duplicated caseIDs}
\usage{
get_duplicate_caseids(duplicates_dir = NULL)
}
\arguments{
\item{duplicates_dir}{directory path where the text files with the duplicates information are.}
}
\value{
an integer vector with all the caseids to be removed
}
\description{
Retrieve the duplicated caseIDs to remove from the analysis.
}
