% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{products_fda}
\alias{products_fda}
\title{List of approved products by FDA}
\format{
A data frame.
}
\source{
<https://www.fda.gov/drugs/drug-approvals-and-databases/drugsfda-data-files>
}
\usage{
products_fda
}
\description{
List of approved products by FDA
}
\keyword{datasets}
