% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_ascii.R
\name{retrieve_faersascii}
\alias{retrieve_faersascii}
\title{Read FAERS ascii files}
\usage{
retrieve_faersascii(
  ascii_dir,
  cache_path = NULL,
  drug_indication_pattern = NULL,
  drug_pattern = NULL,
  primary_suspect = TRUE
)
}
\arguments{
\item{ascii_dir}{directory path where ascii files are}

\item{cache_path}{(optional) a string. Must have a ".Rdata" extension to save the read tabular formats in each loop.}

\item{drug_indication_pattern}{(optional) a string.}

\item{drug_pattern}{(optional) a string.}

\item{primary_suspect}{(optional) a string.}
}
\value{
A list with binded tibbles retrieved from files.
}
\description{
Read FAERS ascii files
}
