% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_xml.R
\name{retrieve_faersxml}
\alias{retrieve_faersxml}
\title{Convert FAERS xml to tabular format}
\usage{
retrieve_faersxml(
  xml_address,
  reaction_wise = TRUE,
  drug_wise = FALSE,
  drug_indication_pattern = NULL
)
}
\arguments{
\item{xml_address}{XML address to be read}

\item{reaction_wise}{each row corresponds to a reaction (if TRUE, drug_wise cannot be TRUE)}

\item{drug_wise}{each row corresponds to a drug (if TRUE, reaction_wise cannot be TRUE)}

\item{drug_indication_pattern}{filter by ADR with a specific drug indication pattern (**stringr** sintax)}
}
\value{
A tibble corresponding to the XML file
}
\description{
Convert FAERS xml to tabular format
}
