% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_xml.R
\name{retrieve_faersxml_all}
\alias{retrieve_faersxml_all}
\title{Convert FAERS a number of xml files to tabular format}
\usage{
retrieve_faersxml_all(xml_address_vector, ..., cache_path = NULL)
}
\arguments{
\item{xml_address_vector}{Vector with XML addresses to be read}

\item{...}{arguments to be passed to \link{retrieve_faersxml}}

\item{cache_path}{a string. Must have a ".Rdata" extension to save the read tabular formats in each loop.}
}
\value{
A binded tibble with all the tibbles returned from `retrieve_faersxml`
}
\description{
Convert FAERS a number of xml files to tabular format
}
