% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_xml.R
\name{retrieve_unique_info}
\alias{retrieve_unique_info}
\title{Retrieve unique drug and ADR information values from XML files}
\usage{
retrieve_unique_info(xml_address_vector, ...)
}
\arguments{
\item{xml_address_vector}{Vector with XML addresses to be read}

\item{...}{arguments to be passed to \link{retrieve_faersxml}}
}
\value{
A list with all the unique information on FAERS variables
}
\description{
Retrieve unique drug and ADR information values from XML files
}
