% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_faersdata}
\alias{summary_faersdata}
\title{FAERS description}
\usage{
summary_faersdata(tabular_faers_data)
}
\arguments{
\item{tabular_faers_data}{a tibble corresponding to the unified FAERS tabular format. Output of function \link{unify_tabular_ascii}}
}
\value{
A list with a findings summary
}
\description{
FAERS description
}
\examples{
summary_faersdata(als_faers_data_unified)
}
