% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabular_faersdata_example}
\alias{tabular_faersdata_example}
\title{Tabular ASCII data example}
\format{
A data frame with 200 rows and 38 columns:
}
\source{
<https://www.fda.gov/drugs/questions-and-answers-fdas-adverse-event-reporting-system-faers/fda-adverse-event-reporting-system-faers-latest-quarterly-data-files>
}
\usage{
tabular_faersdata_example
}
\description{
A subset of data from FAERS data. One row corresponds to one adverse drug reaction.
}
\keyword{datasets}
