% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unified_als_data}
\alias{unified_als_data}
\title{Tabular ASCII data example}
\format{
A data frame with 1635 rows and 40 columns.
}
\usage{
unified_als_data
}
\description{
A subset of data from FAERS data. One row corresponds to one adverse drug reaction. All the ADR in this subset have a primary suspect drug indicated for ALS.
Data frame originated from \link{unify_tabular_ascii}
}
\keyword{datasets}
