% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_func_ascii.R
\name{unify_tabular_ascii}
\alias{unify_tabular_ascii}
\title{Unify the list to a tabular format}
\usage{
unify_tabular_ascii(ascii_list)
}
\arguments{
\item{ascii_list}{list from \link{retrieve_faersascii}}
}
\value{
A data frame representing FAERS data, with all components from the list joined.
}
\description{
Turn the list elements returned from \link{retrieve_faersascii} into a tabular format
}
\examples{
unify_tabular_ascii(ascii_list = als_faers_data)


}
