% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{unzip_faerszip}
\alias{unzip_faerszip}
\title{Unzip FAERS zip folders}
\usage{
unzip_faerszip(zip_folders_dir, ex_dir)
}
\arguments{
\item{zip_folders_dir}{directory containing FAERS zip folders}

\item{ex_dir}{directory to be exported the unzipped files}
}
\value{
None. Just unzips the folders to a specified location.
}
\description{
Unzip FAERS zip folders
}
